+++ tools/web.sh
@@ -1,7 +1,8 @@
 #!/bin/sh -eu
 # takes one arg - directory of a clone of uglymol.github.io
-
 outdir="$1"
+
+[ -e src/elmap.js ] || { echo "Run me from top-level uglymol dir"; exit 1; }
 
 # use README.md without badges
 cat >$outdir/index.md <<EOF
